refactor formatting part
This reduced buffer usage and prevents otherwise unnecessary memory
copies. Downside is that the driver layer gets a bit heavery, but
that's not really an issue now. This will also make it easier to add
an API where the user can provide the buffer (both useful for
embededed systems and large messages).