Remove "--enable-cached-man-pages" switch and make rst2man optional when man pages already exist
liblogging-1.0.2 introduced a new "--enable-cached-man-pages" switch to
allow building the package without the need for rst2man from docutils.

This was a problem, because this was only working with the "special"
release tarball containing pregenerated man pages.
When building from source or after you cleaned the source from the release
tarball, you experienced the following error when you used the
"--enable-cached-man-pages" switch:

	stdlogctl.rst: Command not found

This was because we didn't check for rst2man when the switch was used and
therefore $RST2MAN was undefined when the Makefile tried to build the man
page target which failed.

This patch will remove the switch. Therefore we are now checking if
every man page already exists or not. If a man page is missing we'll check
for rst2man because it is still used in the Makefile (if every man page
exists, make doesn't have to build the target and therefore won't use
the still undefined $RST2MAN variable).

Fixes #13