Issue-32: stdlog should support functionality like the syslog LOG_PID option
Implement a STDLOG_PID option to specify that each log message should be tagged
with the process identifier as well as the ident. This is done in the format
"ident[pid]" as described in RFC-3164. This is honored by both the file: and
uxsock: channel drivers.

Update the test program to allow use of the STDLOG_PID option.